[REFACTOR] 패키지 구조 변경 및 데이터, 뷰모델 클래스 생성
