+++ .idea/misc.xml
@@ -10,7 +10,7 @@
         <entry key="..\:/Users/brian/AndroidStudioProjects/JointSeminarMelon/app/src/main/res/layout/fragment_search.xml" value="0.1" />
         <entry key="..\:/Users/brian/AndroidStudioProjects/JointSeminarMelon/app/src/main/res/layout/fragment_storage.xml" value="0.1" />
         <entry key="..\:/Users/brian/AndroidStudioProjects/JointSeminarMelon/app/src/main/res/menu/menu_bottom_navi.xml" value="0.25364583333333335" />
-        <entry key="..\:/Users/brian/AndroidStudioProjects/Melon-AOS/app/src/main/res/layout/activity_posting.xml" value="0.264" />
+        <entry key="..\:/Users/brian/AndroidStudioProjects/Melon-AOS/app/src/main/res/layout/activity_posting.xml" value="0.1" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/ic_icon_home.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1" />
