+++ app/src/main/java/org/sopt/jointseminar/melon/data/posting/RequestPostingData.kt
@@ -1,0 +1,5 @@
+package org.sopt.jointseminar.melon.data.posting
+
+data class RequestPostingData(
+    val review : String
+)
