+++ app/src/main/java/org/sopt/jointseminar/melon/presentation/posting/PostingViewModel.kt
@@ -1,0 +1,4 @@
+package org.sopt.jointseminar.melon.presentation.posting
+
+class PostingViewModel {
+}
